project/requirements.txt
asgiref==3.7.2
colorama==0.4.6
dj-rest-auth==4.0.1
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.7
exceptiongroup==1.1.2
inflection==0.5.1
iniconfig==2.0.0
mysqlclient==2.2.0
packaging==23.1
pluggy==1.2.0
PyJWT==1.7.1
PyMySQL==1.1.0
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
